Add front view on React/Redux

+++ src/main/java/com/little_pig_bot/repository/NoteRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.little_pig_bot.model.Note;
 
 @Repository
-public interface NoteRepository extends JpaRepository<Note, Integer> {
+public interface NoteRepository extends JpaRepository<Note, Long> {
 
     List<Note> findAllByUserId(Integer userId);
